Create training controllers, services, repositories. Build the appearance of a email letter to match the style of the site

+++ src/main/java/by/yankavets/typingtrainer/exception/user/UserNotFoundException.java
@@ -4,7 +4,7 @@ public class UserNotFoundException extends RuntimeException {
 
     private static final String MESSAGE_WITH_ID = "User with ID %d not found";
 
-    public UserNotFoundException(Integer id) {
+    public UserNotFoundException(long id) {
         super(String.format(MESSAGE_WITH_ID, id));
     }
 }
